requirements.txt
appdirs==1.4.4
asgiref==3.3.1
astroid==2.5
botocore==1.20.43
certifi==2020.12.5
chardet==4.0.0
cloudinary==1.25.0
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.7
django-cloudinary-storage==0.3.0
djangorestframework==3.12.4
filelock==3.0.12
idna==2.10
isort==5.7.0
jmespath==0.10.0
lazy-object-proxy==1.5.2
mccabe==0.6.1
Pillow==8.2.0
psycopg2-binary==2.8.6
pylint==2.7.1
pylint-plugin-utils==0.6
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
requests==2.25.1
s3transfer==0.3.6
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.4
virtualenv==20.4.3
wrapt==1.12.1
